bot/requirements.txt
Pillow==8.3.0
pyzbar==0.1.8
psycopg2-binary==2.9.1
pyTelegramBotAPI==3.8.1
python-decouple==3.3
redis==3.5.3